Refactor e2e files
