+++ e2e/tests/page.spec.ts
@@ -6,8 +6,6 @@ import { Dayjs } from "dayjs";
 
 const mockEntries = getMockEntries();
 
-console.log(mockEntries);
-
 test.beforeEach(async ({ page }) => {
   await page.goto("/");
 });
@@ -40,51 +38,14 @@ test.describe("Landing page", () => {
         .minute(Number(entry.hours) * 60)
         .minute();
       const hour = dayjs().hour(Number(entry.hours)).hour();
-      console.log(minute, hour);
       let entryRows = workdayEntryList.getByRole("listitem");
       // Filter row by field texts
       for (const field of entry.fields) {
         entryRows = entryRows.filter({ hasText: new RegExp(`${field.DimensionItem}`) });
       }
       entryRows = entryRows.filter({ hasText: new RegExp(`${hour}:${minute}`) });
-      await expect(entryRows).toBeVisible();
+      await expect(entryRows.first()).toBeVisible();
     }
-  });
-});
-
-test.describe("Browse week", () => {
-  const startingWeek = 21;
-  test.beforeEach(async ({ page }) => {
-    await page.goto(`/entries/week/${startingWeek}`);
-  });
-
-  test("Should browse weeks forward", async ({ page, t }) => {
-    const jump = 3;
-    for (let i = 0; i < jump; i++)
-      await page.getByRole("button", { name: t("controls.aria.nextWeek") }).click();
-    await expect(page).toHaveURL(`/entries/week/${startingWeek + jump}`);
-  });
-
-  test("Should browse weeks backward", async ({ page, t }) => {
-    const jump = 3;
-    for (let i = 0; i < jump; i++)
-      await page.getByRole("button", { name: t("controls.aria.prevWeek") }).click();
-    await expect(page).toHaveURL(`/entries/week/${startingWeek - jump}`);
-  });
-
-  test("Should go to current week", async ({ page, t, dayjs }) => {
-    const jump = 10;
-    const currentWeek = dayjs().week();
-    await page.goto(`/entries/week/${currentWeek + jump}`);
-    await page.getByRole("button", { name: t("controls.aria.currentWeek") }).click();
-    await expect(page).toHaveURL(`/entries/week/${currentWeek}`);
-  });
-
-  test("Should go to specific week", async ({ page, t }) => {
-    const jump = 4;
-    await page.getByRole("combobox", { name: String(startingWeek) }).click();
-    await page.getByRole("option", { name: String(startingWeek + jump) }).click();
-    await expect(page).toHaveURL(`/entries/week/${startingWeek + jump}`);
   });
 });
